Add new hyperscaler file
