+++ internal/clusterconfig/hyperscaler.go
@@ -1,0 +1,5 @@
+package clusterconfig
+
+func IsHyperscalerAWS() (bool, error) {
+	return true, nil
+}
